+++ R/cplot_polr.R
@@ -66,13 +66,6 @@ function(object,
 
     # setup `outdat` data
     if (what %in% c("prediction", "classprediction", "stackedprediction")) {
-        mean_or_mode <- function(x) {
-            if (is.factor(x)) {
-                factor(names(sort(table(x), descending = TRUE))[1L], levels = levels(x))
-            } else {
-                mean(x, na.rm = TRUE)
-            }
-        }
         tmpdat <- lapply(dat[, names(dat) != xvar, drop = FALSE], mean_or_mode)
         tmpdat <- structure(lapply(tmpdat, rep, length(xvals)),
                             class = "data.frame", row.names = seq_len(length(xvals)))